OTA_GH/client/requirements.txt
# OTA Client Dependencies

requests==2.31.0
paho-mqtt==1.6.1
python-dotenv==1.0.0

# LLM 2차 검증 - 시스템 리소스 수집
psutil>=5.9.0